split to new file
