+++ server/src/test/groovy/com/exacaster/lighter/application/sessions/SessionServiceTest.groovy
@@ -1,6 +1,5 @@
 package com.exacaster.lighter.application.sessions
 
-import com.exacaster.lighter.application.ApplicationInfo
 import com.exacaster.lighter.application.ApplicationState
 import com.exacaster.lighter.application.sessions.processors.StatementHandler
 import com.exacaster.lighter.backend.Backend
@@ -11,7 +10,7 @@ import com.exacaster.lighter.test.InMemoryStorage
 import spock.lang.Specification
 import spock.lang.Subject
 
-import static com.exacaster.lighter.test.Factories.*
+import static com.exacaster.lighter.test.Factories.submitParams
 
 class SessionServiceTest extends Specification {
     ApplicationStorage storage = new InMemoryStorage()
@@ -68,48 +67,4 @@ class SessionServiceTest extends Specification {
         then: "returns empty"
         session.isEmpty()
     }
-
-    def "create statement"() {
-        given:
-        def params = newStatement()
-
-        when: "creating statement"
-        def result = service.createStatement("sessionId", params)
-
-        then: "returns no session found"
-        result instanceof StatementCreationResult.NoSessionExists
-    }
-
-    def "create statement2"() {
-        given:
-        def params = newStatement()
-        def session = newSession(ApplicationState.KILLED)
-        storage.saveApplication(session)
-        backend.getInfo(session) >>  Optional.of( new ApplicationInfo(session.state, session.id))
-
-        when: "creating statement"
-        def result = service.createStatement(session.id, params)
-
-        then: "returns session in invalid state"
-        result instanceof StatementCreationResult.SessionInInvalidState
-        ((StatementCreationResult.SessionInInvalidState)result).getInvalidState() == session.state
-    }
-
-    def "create statement3"() {
-        given:
-        def statementToCreate = newStatement()
-        def session = newSession(ApplicationState.STARTING)
-        storage.saveApplication(session)
-        backend.getInfo(session) >>  Optional.of( new ApplicationInfo(session.state, session.id))
-        def statementCreated = statement()
-        statementHandler.processStatement(session.id, statementToCreate) >> statementCreated
-
-        when: "creating statement"
-        def result = service.createStatement(session.id, statementToCreate)
-
-        then: "returns statement created"
-        result instanceof StatementCreationResult.StatementCreated
-        ((StatementCreationResult.StatementCreated)result).getStatement() == statementCreated
-    }
-
 }
